src/requirements.txt
pandas
biopython
dnachisel
jupyterlab
scikit-bio
pytest
matplotlib
